+++ illumio-core/provider_test.go
@@ -10,6 +10,7 @@ import (
 	"testing"
 
 	"github.com/hashicorp/go-version"
+	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/resource"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/terraform"
 )
@@ -92,5 +93,60 @@ func skipIfPCEVersionBelow(v string) func() (bool, error) {
 		}
 
 		return pceVersion.LessThan(checkVersion), nil
+	}
+}
+
+// deleteFromPCE removes the label with the given HREF
+// via the API to test Terraform drift behaviour
+func deleteFromPCE(href *string, t *testing.T) func() {
+	return func() {
+		conf := TestAccProvider.Meta().(Config)
+		illumioClient := conf.IllumioClient
+
+		_, err := illumioClient.Delete(*href)
+		if err != nil {
+			t.Fatal("Failed to delete object with HREF " + *href + " from PCE")
+		}
+	}
+}
+
+// testAccCheckResourceExists checks if a resource exists and assigns
+// the corresponding HREF to the given pointer
+func testAccCheckResourceExists(n string, href *string) resource.TestCheckFunc {
+	return func(s *terraform.State) error {
+		// find the corresponding state object
+		rs, ok := s.RootModule().Resources[n]
+		if !ok {
+			return fmt.Errorf("Not found: %s", n)
+		}
+
+		// retrieve the configured client from the test setup
+		conf := TestAccProvider.Meta().(Config)
+		illumioClient := conf.IllumioClient
+
+		_, data, err := illumioClient.Get(rs.Primary.ID, nil)
+		if err != nil {
+			return err
+		}
+
+		ref := data.S("href").Data().(string)
+		*href = ref
+
+		return nil
+	}
+}
+
+// testAccCheckCompareRefs compares given HREFs
+// XXX: note that we have to pass the HREFs as pointers
+// as this function is run before the tests; the inner
+// function needs references to the refs that can be set
+// during the test
+func testAccCheckCompareRefs(origRef, newRef *string, shouldEqual bool) resource.TestCheckFunc {
+	return func(s *terraform.State) error {
+		if (origRef == newRef) == shouldEqual {
+			return nil
+		}
+
+		return fmt.Errorf("testAccCheckCompareRefs(%q, %q, %t) failed", *origRef, *newRef, shouldEqual)
 	}
 }